\name{mod_info}
\alias{mod_info}
\title{This is a function to prepare modification information data set}
\usage{
mod_info <- function(RpsiScan_res_file,
                     RpsiAnnotator_res_file,
                     psiU.SingleSites.bed,
                     pred_method,
                     output_dir,
                     output_name)
}
\arguments{
  \item{RpsiScan_res_file}{The psiScan result file generated by RpsiScan function.}
  \item{RpsiAnnotator_res_file}{The psiAnnotator result file generated by RpsiAnnotator function.}
  \item{psiU.SingleSites.bed}{pseudouridylation sites in bed format, default we recommend hg38.psiU.SingleSites.bed.}
  \item{pred_method}{prediction method, current version provided option: 'roc', 'user-defined','svm','ann'.}
  \item{output_dir}{Directory of the output file.}
  \item{output_name}{Name of the output file.}
}
\description{
Prepare modification information data set (known pseudouridylation sites) for pseudouridylation identification.
}
\examples{

mod_info(RpsiScan_res_file="../test_out/roc/roc_roc_psi_prediction.bed",
         RpsiAnnotator_res_file="../test_out/roc_annotation/roc_annotation_anno.bed",
         psiU.SingleSites.bed="../test_data/ground_truth/hg38.psiU.SingleSites.bed",
         pred_method="roc",
         output_dir="../test_out/mod_info",
         output_name="mod_info")

mod_info(RpsiScan_res_file="../test_out/user_defined/user_defined_ud_psi_prediction.bed",
         RpsiAnnotator_res_file="../test_out/ud_annotation/ud_annotation_anno.bed",
         psiU.SingleSites.bed="../test_data/ground_truth/hg38.psiU.SingleSites.bed",
         pred_method="user-defined",
         output_dir="../test_out/mod_info",
         output_name="mod_info")

mod_info(RpsiScan_res_file="../test_out/svm_predict/svm_predict_svm_psi_prediction.bed",
         RpsiAnnotator_res_file="../test_out/svm_annotation/svm_annotation_anno.bed",
         psiU.SingleSites.bed="../test_data/ground_truth/hg38.psiU.SingleSites.bed",
         pred_method="svm",
         output_dir="../test_out/mod_info",
         output_name="mod_info")

mod_info(RpsiScan_res_file="../test_out/ann_predict/ann_predict_ann_psi_prediction.bed",
         RpsiAnnotator_res_file="../test_out/ann_annotation/ann_annotation_anno.bed",
         psiU.SingleSites.bed="../test_data/ground_truth/hg38.psiU.SingleSites.bed",
         pred_method="ann",
         output_dir="../test_out/mod_info",
         output_name="mod_info")


}
\value{
tab-delimited text files include modification information (based on known pseudouridylation sites)}
