\name{psiSVM}
\alias{psiSVM}
\title{Build SVM model based on ground truth pseudouridylation data set}
\usage{
psiSVM <- function(rocfile,
                  rRNAfile,
                  rRNAfile2,
                  filtfile,
                  output_dir,
                  output_name)
}
\arguments{
  \item{rocfile}{SVM input file of single sites information (normally with suffix _roc_plot.txt) generated from ground_truth() function.}
  \item{rRNAfile}{The bed file contains knonw rRNA pseudouridylation sites, default we recommend hg38_human_chr21_rRNA_known_pseudoU_SingleSites.bed.}
  \item{rRNAfile2}{default we recommend hg38.psiU.SingleSites.bed.}
  \item{filtfile}{File to be filted.}
  \item{output_dir}{Directory of the output file.}
  \item{output_name}{Name of the output file.}
}
\description{This function will accept result file generated by ground_truth() function and give out evaluation on input ground truth rRNA pseudouridylation data set.}
\examples{
RpsiScan_example()#when finished, view your global environment, and you will detect Data named 'RpsiScan_example'

prefilt_out <- prefilt(RpsiScan_res=RpsiScan_example,
                        target_chrom="^chr[0-9|a-z|A-Z]*$",
                        target_base="T",
                        output_dir="../test_out/RpsiScan_out",
                        output_name="test_prefilt")

ground_truth(RpsiScan_res_file="../test_out/RpsiScan_out/test_prefilt.txt",
              known_rRNA_bed="../test_data/ground_truth/hg38_human_chr21_rRNA_known_pseudoU_SingleSites.bed",
              rrna_chr21_bed="../test_data/ground_truth/rrna_chr21.bed",
              output_dir="../test_out/ground_truth",
              output_name="ground_truth")

psiSVM(rocfile="../test_out/ground_truth/ground_truth_roc_plot.txt",
        rRNAfile="../test_data/ground_truth/hg38_human_chr21_rRNA_known_pseudoU_SingleSites.bed",
        rRNAfile2="../test_data/ground_truth/hg38.psiU.SingleSites.bed",
        filtfile="../test_out/RpsiScan_out/test_prefilt.txt",
        output_dir="../test_out/svm_model",
        output_name="svm_model")

}
\value{Building SVM model based on e1071}
