\name{RpsiAnnotator}
\alias{RpsiAnnotator}
\title{R API for psiAnnotator}
\usage{
RpsiAnnotator <- function(input_bedfile,
                         annotation_bed6file,
                         output_dir,
                         output_name)
}
\arguments{
  \item{input_bedfile}{The input bed file to be annotated.}
  \item{annotation_bed6file}{The bed6 file for annotation, default we recommend hg38.genecode.v30.tRNA.snoRNA.miRNA.rmsk.exonFeatures.bed6.}
  \item{output_dir}{Directory of the output file.}
  \item{output_name}{Name of the output file.}
  }
\description{Run psiAnnotator for pseudouridylation annotation.}
\details{
If error was showed like: 'RpsiScan/program/psiAnnotator: error while loading shared libraries: libssl.so.10: cannot open shared object file: No such file or directory', we recommend fixing the error by the following steps (psiAnnotator running require libcrypto.so.10/libpng12.so.0/libssl.so.10):
1. cd /the/directory/of/RpsiScan/inst/program/
2. sudo cp lib* /lib/x86_64-linux-gnu
}
\examples{

#annotate psiScan result generated by User-defined thresholds
RpsiAnnotator(input_bedfile="../test_out/RpsiScan_out/test_udfilt.bed",
            annotation_bed6file="../test_data/annotation/hg38.genecode.v30.tRNA.snoRNA.miRNA.rmsk.exonFeatures.bed6",
            output_dir="../test_out/ud_annotation",
            output_name = "ud_annotation")

#annotate prediction result generated by ROC best threshold
RpsiAnnotator(input_bedfile="../test_out/roc/roc_roc_psi_prediction.bed",
            annotation_bed6file="../test_data/annotation/hg38.genecode.v30.tRNA.snoRNA.miRNA.rmsk.exonFeatures.bed6",
            output_dir="../test_out/roc_annotation",
            output_name = "roc_annotation")

#annotate prediction result generated by SVM model
RpsiAnnotator(input_bedfile="../test_out/svm_predict/svm_predict_svm_psi_prediction.bed",
            annotation_bed6file="../test_data/annotation/hg38.genecode.v30.tRNA.snoRNA.miRNA.rmsk.exonFeatures.bed6",
            output_dir="../test_out/svm_annotation",
            output_name = "svm_annotation")

#annotate prediction result generated by ANN model
RpsiAnnotator(input_bedfile="../test_out/ann_predict/ann_predict_ann_psi_prediction.bed",
            annotation_bed6file="../test_data/annotation/hg38.genecode.v30.tRNA.snoRNA.miRNA.rmsk.exonFeatures.bed6",
            output_dir="../test_out/ann_annotation",
            output_name = "ann_annotation")
}
\value{RpsiAnnotator will run psiAnnotator to add annotation information for each single base.}
