\name{RpsiScan_example}
\alias{RpsiScan_example}
\title{Get RpsiScan example result}
\usage{RpsiScan_example <- function()}
\description{This function will get an example psiFinder result from CIV-seq sequencing data.}
\examples{RpsiScan_example()#when finished, view your global environment, and you will detect Data named 'RpsiScan_example'}
\value{tab-delimited text files include stop/mutation/deletion read information for each treatment and control group, e.g. tretRtsRatio, ctrlRtsRatio, tretMutRatio, ctrlMutRatio, tretDelRatio, ctrlDelRatio.}
