\name{user_defined}
\alias{user_defined}
\title{Identify pseudouridylation sites based on user-defined thresholds}
\usage{
user_defined <- function(rocfile,
                        rRNAfile,
                        filtfile,
                        output_dir,
                        output_name)
}
\arguments{
  \item{rocfile}{User-defined input file of single sites information (normally with suffix _roc_plot.txt) generated from ground_truth() function.}
  \item{rRNAfile}{The bed file contains knonw rRNA pseudouridylation sites, default we recommend hg38_human_chr21_rRNA_known_pseudoU_SingleSites.bed.}
  \item{filtfile}{File to be filted.}
  \item{output_dir}{Directory of the output file.}
  \item{output_name}{Name of the output file.}
}
\description{This function will accept result file generated by ground_truth() function and give out evaluation on input ground truth rRNA pseudouridylation data set.}
\examples{
RpsiScan_example()#when finished, view your global environment, and you will detect Data named 'RpsiScan_example'

prefilt_out <- prefilt(RpsiScan_res=RpsiScan_example,
                        target_chrom="^chr[0-9|a-z|A-Z]*$",
                        target_base="T",
                        output_dir="../test_out/RpsiScan_out",
                        output_name="test_prefilt")

udfilt_out <- udfilt(RpsiScan_res_file="../test_out/RpsiScan_out/test_prefilt.txt",
                   tretRtsRatio_thres=0.02,
                   rtsRatioFold_thres=1.5,
                   tretBefRatio_thres=0.02,
                   befRatioFold_thres=1.5,
                   tretAftRatio_thres=0.02,
                   aftRatioFold_thres=1.5,
                   tretMutRatio_thres=0.02,
                   mutRatioFold_thres=1.5,
                   tretDelRatio_thres=0.02,
                   delRatioFold_thres=1.5,
                   output_dir="../test_out/RpsiScan_out",
                   output_name="test_udfilt")

ground_truth(RpsiScan_res_file="../test_out/RpsiScan_out/test_udfilt.txt",
              known_rRNA_bed="../test_data/ground_truth/hg38_human_chr21_rRNA_known_pseudoU_SingleSites.bed",
              rrna_chr21_bed="../test_data/ground_truth/rrna_chr21.bed",
              output_dir="../test_out/ground_truth",
              output_name="ground_truth")

user_defined(rocfile="../test_out/ground_truth/ground_truth_roc_plot.txt",
              rRNAfile="../test_data/ground_truth/hg38_human_chr21_rRNA_known_pseudoU_SingleSites.bed",
              filtfile="../test_out/RpsiScan_out/test_udfilt.txt",
              output_dir="../test_out/user_defined",
              output_name="user_defined")

}
\value{Evaluation result (figures and tables) contain TPR, TNR, etc.}
